fix: add casting

+++ 2_module/ex01/Fixed.cpp
@@ -17,7 +17,7 @@ Fixed::Fixed(const int integer): rawBits(integer << fractionalBits)
     std::cout << INT_CONSTRUCTOR << std::endl;
 }
 
-Fixed::Fixed(const float float_nbr): rawBits(roundf(float_nbr * (1 << fractionalBits)))
+Fixed::Fixed(const float float_nbr): rawBits(roundf(float_nbr * (float )(1 << fractionalBits)))
 {
     std::cout << FLOAT_CONSTRUCTOR << std::endl;
 }
